+++ envdoctor/__init__.py
@@ -1,4 +1,0 @@
-"""envdoctor explains why Python environment config is broken."""
-
-__version__ = "0.1.0"
-