+++ envdoctor/reporters/__init__.py
@@ -1,2 +1,0 @@
-"""Render envdoctor check results."""
-