+++ envgap/model/__init__.py
@@ -1,0 +1,13 @@
+from envgap.model.env_source import EnvFile, EnvVar
+from envgap.model.expected_var import CodeUsage, ExpectedVar
+from envgap.model.finding import Finding, Severity
+
+__all__ = [
+    "CodeUsage",
+    "EnvFile",
+    "EnvVar",
+    "ExpectedVar",
+    "Finding",
+    "Severity",
+]
+