+++ envgap/reporters/__init__.py
@@ -1,0 +1,2 @@
+"""Render envgap check results."""
+